refactor: reduce duplicated page and artist UI
- Extract PageShell, PageHeader, ResponsiveContainer primitives; apply across all pages
- Extract ArtistCard component (home + featured variants) to consolidate artist card rendering
- Render game result/instruction card from one sideCardContent variable in DynamicHeardle
- Split DailyChallengeStatus into its own file; type custom event detail; delete unused ArtistHeader export
