+++ .gitignore
@@ -48,3 +48,6 @@ next-env.d.ts
 
 #keys
 /keys
+
+# Temporary files
+.tmp-*
