+++ src/app/[artist]/page.tsx
@@ -7,10 +7,10 @@ import DynamicHeardle from '@/components/game/DynamicHeardle';
 import ModeSelector from '@/components/game/ModeSelector';
 import StatisticsButton from '@/components/stats/StatisticsButton';
 import PageLoadingSpinner from '@/components/ui/LoadingSpinner';
-import { DailyChallengeStatus } from '@/components/artist/ArtistHeader';
+import { DailyChallengeStatus } from '@/components/artist/DailyChallengeStatus';
 import NextDailyCountdown from '@/components/game/NextDailyCountdown';
-import AnimatedBackground from '@/components/ui/AnimatedBackground';
 import NotificationBanner from '@/components/ui/NotificationBanner';
+import { PageShell, PageHeader, ResponsiveContainer } from '@/components/ui/PageShell';
 import RelatedArtists from '@/components/artist/RelatedArtists';
 import FAQSection from '@/components/seo/FAQSection';
 import Breadcrumbs from '@/components/ui/Breadcrumbs';
@@ -45,98 +45,85 @@ export default function ArtistPage({ params }: ArtistPageProps) {
 
   if (!artist) {
     return (
-      <div className="min-h-screen bg-gradient-to-br from-slate-900 via-purple-900 to-slate-900 flex items-center justify-center relative overflow-hidden">
-        <AnimatedBackground blobCount={2} subtle />
-        <div className="relative z-10 text-center">
-          <div className="backdrop-blur-xl bg-white/10 rounded-3xl border border-white/20 p-12">
-            <h1 className="text-5xl font-bold text-white mb-6">Artist Not Found</h1>
-            <p className="text-white/80 mb-8 text-lg">The artist you're looking for doesn't exist.</p>
-            <a
-              href="/"
-              className="inline-block px-8 py-4 bg-gradient-to-r from-pink-500 to-purple-600 text-white font-bold rounded-2xl hover:shadow-2xl hover:shadow-purple-500/25 transition-all duration-300 transform hover:scale-105"
-            >
-              ← Back to Home
-            </a>
+      <PageShell blobCount={2} subtle>
+        <div className="flex items-center justify-center min-h-screen">
+          <div className="relative z-10 text-center">
+            <div className="backdrop-blur-xl bg-white/10 rounded-3xl border border-white/20 p-12">
+              <h1 className="text-5xl font-bold text-white mb-6">Artist Not Found</h1>
+              <p className="text-white/80 mb-8 text-lg">The artist you're looking for doesn't exist.</p>
+              <a
+                href="/"
+                className="inline-block px-8 py-4 bg-gradient-to-r from-pink-500 to-purple-600 text-white font-bold rounded-2xl hover:shadow-2xl hover:shadow-purple-500/25 transition-all duration-300 transform hover:scale-105"
+              >
+                ← Back to Home
+              </a>
+            </div>
           </div>
         </div>
-      </div>
+      </PageShell>
     );
   }
 
   return (
-    <div className="min-h-screen bg-gradient-to-br from-slate-900 via-purple-900 to-slate-900 relative overflow-hidden">
-      <AnimatedBackground blobCount={3} subtle />
+    <PageShell blobCount={3} subtle>
+      <PageHeader>
+        <div className="py-3 sm:py-4">
+          {/* Top row: Breadcrumbs, Title, Stats */}
+          <div className="flex justify-between items-center mb-2">
+            <div className="flex-shrink-0">
+              <Breadcrumbs
+                items={[
+                  { label: 'Home', href: '/' },
+                  { label: artist.displayName }
+                ]}
+              />
+            </div>
 
-      {/* Header with integrated content */}
-       <div className="relative z-10 backdrop-blur-md bg-white/10 border-b border-white/20">
-         <div className="w-full px-4 sm:px-6 lg:px-8 xl:px-12 2xl:px-16">
-           <div className="py-3 sm:py-4">
-             {/* Top row: Breadcrumbs, Title, and Stats */}
-             <div className="flex justify-between items-center mb-2">
-               {/* Left: Breadcrumbs */}
-               <div className="flex-shrink-0">
-                 <Breadcrumbs 
-                   items={[
-                     { label: 'Home', href: '/' },
-                     { label: artist.displayName }
-                   ]}
-                 />
-               </div>
-               
-               {/* Center: Artist Title */}
-               <div className="absolute left-1/2 transform -translate-x-1/2">
-                 <h1 className="text-xl sm:text-2xl md:text-3xl font-bold text-white whitespace-nowrap">
-                   {artist.displayName} Heardle
-                 </h1>
-               </div>
-               
-               {/* Right: Stats Button */}
-               <div className="flex-shrink-0">
-                 <StatisticsButton artistId={artist.id} currentMode={selectedMode} />
-               </div>
-             </div>
-             
-             {/* Description row */}
-             <div className="text-center">
-               <p className="text-white/70 text-xs sm:text-sm md:text-base max-w-2xl mx-auto leading-relaxed">
-                 {artist.fandom ? `Calling all ${artist.fandom}! ` : ''}
-                 Test your knowledge of {artist.displayName}'s discography. 
-                 Listen to song clips and guess the title. New daily challenge every day!
-               </p>
-             </div>
-           </div>
-         </div>
-       </div>
+            <div className="absolute left-1/2 transform -translate-x-1/2">
+              <h1 className="text-xl sm:text-2xl md:text-3xl font-bold text-white whitespace-nowrap">
+                {artist.displayName} Heardle
+              </h1>
+            </div>
 
-      {/* Notification Banner */}
+            <div className="flex-shrink-0">
+              <StatisticsButton artistId={artist.id} currentMode={selectedMode} />
+            </div>
+          </div>
+
+          <div className="text-center">
+            <p className="text-white/70 text-xs sm:text-sm md:text-base max-w-2xl mx-auto leading-relaxed">
+              {artist.fandom ? `Calling all ${artist.fandom}! ` : ''}
+              Test your knowledge of {artist.displayName}&apos;s discography.
+              Listen to song clips and guess the title. New daily challenge every day!
+            </p>
+          </div>
+        </div>
+      </PageHeader>
+
       <NotificationBanner
         id="new-artists-2026-02"
         message="🎵 New Update! We've added over 100 K-pop artists. Explore more on the home page!"
         icon="✨"
       />
 
-             {/* Main Content */}
-       <div className="relative z-10 w-full px-4 sm:px-6 lg:px-8 xl:px-12 2xl:px-16 py-4">
-         <div className="flex flex-col items-stretch space-y-4 sm:space-y-5 lg:space-y-6">
-           <ModeSelector 
-            selectedMode={selectedMode} 
-            onModeChange={setSelectedMode} 
+      <ResponsiveContainer className="py-4">
+        <div className="flex flex-col items-stretch space-y-4 sm:space-y-5 lg:space-y-6">
+          <ModeSelector
+            selectedMode={selectedMode}
+            onModeChange={setSelectedMode}
           />
-          
-          {/* Daily Countdown + Daily Challenge Status - Only show in daily mode */}
+
           {selectedMode === 'daily' && (
             <div className="flex flex-col items-center gap-2 sm:gap-3 lg:gap-4">
               <DailyChallengeStatus artistId={artist.id} />
               <NextDailyCountdown />
             </div>
           )}
-          
-          {/* Remove any extra top margin/padding from the child; parent controls spacing */}
+
           <DynamicHeardle
             key={selectedMode}
-            mode={selectedMode} 
+            mode={selectedMode}
             onGameStateChange={(gameState) => {
-              // Force update of the DailyChallengeStatus component
               if (gameState.isGameOver) {
                 const event = new CustomEvent(DAILY_CHALLENGE_UPDATED_EVENT, {
                   detail: { artistId: artist.id, date: getTodayString(), completed: gameState.isGameOver }
@@ -145,14 +132,11 @@ export default function ArtistPage({ params }: ArtistPageProps) {
               }
             }}
           />
-         </div>
-      </div>
+        </div>
+      </ResponsiveContainer>
 
-      {/* Related Artists Section - Internal Linking for SEO */}
       <RelatedArtists currentArtistId={artist.id} maxArtists={4} />
-
-      {/* FAQ Section - UI only (schema rendered server-side in layout) */}
       <FAQSection faqItems={generateFAQItems(artist.displayName, artist.fandom)} />
-    </div>
+    </PageShell>
   );
 }
